Added Firewall Rules WebServer will now attempt to add firewall rules as well as port reservations if the application is running under a user with admin privileges WebServer will no longer attempt to add post reservations if the application is not running under a user with admin privileges WebServer will display the netsh commands required to add and the above when not running under a user with admin privileges

+++ Thingy.WebServerLite/WebServer.cs
@@ -16,6 +16,7 @@ namespace Thingy.WebServerLite
         private readonly IWebServerRequestFactory webServerRequestFactory;
         private readonly IWebServerResponseFactory webServerResponseFactory;
         private readonly IWebServerLoggingProvider logger;
+        private readonly bool isAdmin;
 
         HttpListener listener = null;
 
@@ -29,6 +30,8 @@ namespace Thingy.WebServerLite
             this.webServerRequestFactory = webServerRequestFactory;
             this.webServerResponseFactory = webServerResponseFactory;
             this.logger = logger;
+            this.isAdmin = new WindowsPrincipal(WindowsIdentity.GetCurrent()).IsInRole(WindowsBuiltInRole.Administrator);
+
         }
 
         public void Dispose()
@@ -60,7 +63,23 @@ namespace Thingy.WebServerLite
             {
                 string prefix = string.Format("http://*:{0}/", webSite.PortNumber);
                 listener.Prefixes.Add(prefix);
-                AddHttpNameSpaceReservation(prefix);
+
+                if (isAdmin)
+                {
+                    AddHttpNameSpaceReservation(prefix);
+                    AddFirewallRule(webSite.Name, webSite.PortNumber);
+                }
+                else
+                {
+                    logger.WriteMessage("You are not running in administrator mode. To get the web server to work you may need to enter the following commands in a command window with elevated privileges.");
+                    logger.WriteMessage(GetAddNameSpaceReservationCommand(prefix));
+                    logger.WriteMessage(GetAddFirewallRuleInCommand(webSite.Name, webSite.PortNumber));
+                    logger.WriteMessage(GetAddFirewallRuleOutCommand(webSite.Name, webSite.PortNumber));
+                    logger.WriteMessage("To remove these setttings:");
+                    logger.WriteMessage(GetDeleteNameSpaceReservationCommand(prefix));
+                    logger.WriteMessage(GetRemoveFirewallRuleInCommand(webSite.Name, webSite.PortNumber));
+                    logger.WriteMessage(GetRemoveFirewallRuleOutCommand(webSite.Name, webSite.PortNumber));
+                }
             }
 
             listener.Start();
@@ -70,24 +89,67 @@ namespace Thingy.WebServerLite
 
         private void DeleteHttpNameSpaceReservations()
         {
-            if (webSites != null)
+            if (isAdmin && webSites != null)
             {
                 foreach (IWebSite webSite in webSites)
                 {
                     string prefix = string.Format("http://*:{0}/", webSite.PortNumber);
                     DeleteHttpNameSpaceReservation(prefix);
+                    RemoveFireWallRule(webSite.Name, webSite.PortNumber);
                 }
             }
         }
 
         private void AddHttpNameSpaceReservation(string prefix)
         {
-            RunProcess(string.Format("netsh http add urlacl url={0} user={1}", prefix, WindowsIdentity.GetCurrent().Name));
+            RunProcess(GetAddNameSpaceReservationCommand(prefix));
+        }
+
+        private static string GetAddNameSpaceReservationCommand(string prefix)
+        {
+            return string.Format("netsh http add urlacl url={0} user={1}", prefix, WindowsIdentity.GetCurrent().Name);
         }
 
         private void DeleteHttpNameSpaceReservation(string prefix)
         {
-            RunProcess(string.Format("netsh http delete urlacl url={0}", prefix));
+            RunProcess(GetDeleteNameSpaceReservationCommand(prefix));
+        }
+
+        private static string GetDeleteNameSpaceReservationCommand(string prefix)
+        {
+            return string.Format("netsh http delete urlacl url={0}", prefix);
+        }
+
+        private void AddFirewallRule(string name, int portNumber)
+        {
+            RunProcess(GetAddFirewallRuleInCommand(name, portNumber));
+            RunProcess(GetAddFirewallRuleOutCommand(name, portNumber));
+        }
+
+        private static string GetAddFirewallRuleOutCommand(string name, int portNumber)
+        {
+            return string.Format("netsh advfirewall firewall add rule name=\"Game Server rule : {0}, open port {1}\" dir=out action=allow protocol=TCP localport={1}", name, portNumber);
+        }
+
+        private static string GetAddFirewallRuleInCommand(string name, int portNumber)
+        {
+            return string.Format("netsh advfirewall firewall add rule name=\"Game Server rule : {0}, open port {1}\" dir=in action=allow protocol=TCP localport={1}", name, portNumber);
+        }
+
+        private void RemoveFireWallRule(string name, int portNumber)
+        {
+            RunProcess(GetRemoveFirewallRuleInCommand(name, portNumber));
+            RunProcess(GetRemoveFirewallRuleOutCommand(name, portNumber));
+        }
+
+        private static string GetRemoveFirewallRuleOutCommand(string name, int portNumber)
+        {
+            return string.Format("netsh advfirewall firewall delete rule name=\"Game Server rule : {0}, open port {1}\" dir=out protocol=TCP localport={1}", name, portNumber);
+        }
+
+        private static string GetRemoveFirewallRuleInCommand(string name, int portNumber)
+        {
+            return string.Format("netsh advfirewall firewall delete rule name=\"Game Server rule : {0}, open port {1}\" dir=in protocol=TCP localport={1}", name, portNumber);
         }
 
         private void RunProcess(string command)
